Fixed compilation on Windows

+++ actiontools/QHotkey/qhotkey_win.cpp
@@ -203,7 +203,7 @@ quint32 QHotkeyPrivateWin::nativeKeycode(Qt::Key keycode, bool &ok)
 
 	default:
 		if(keycode <= 0xFFFF)
-			return (byte)keycode;
+            return static_cast<int>(keycode);
 		else {
 			ok = false;
 			return 0;
